services/agent-service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
google-cloud-aiplatform>=1.30.0
google-cloud-storage>=2.7.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.24.0
pytz>=2023.3
 